Hp, int->float형 변경

+++ Arknight/Assets/Scripts/MainScene/Enemy/MonsterStat.cs
@@ -1,14 +1,16 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.UI;
+
 
 public class MonsterStat : MonoBehaviour
 {
-    public int MonsterAttack=70;
-    public int BossAttack = 10;
+    public float MonsterAttack= 70.0f;
+    public float BossAttack = 10.0f;
     public float AttackDelay = 2.0f;
-    public int MaxHp = 233;
-    public int CurrentHP = 0;
+    public float MaxHp = 233.0f;
+    public float CurrentHP = 0.0f;
 
     // Start is called before the first frame update
     void Start()
@@ -16,14 +18,14 @@ public class MonsterStat : MonoBehaviour
         CurrentHP = MaxHp;
 
     }
-    public bool UpdateHP(int v)
+    public bool UpdateHP(float v)
     {
         CurrentHP += v;
         if (CurrentHP > MaxHp) CurrentHP = MaxHp;
-        if (CurrentHP <= 0)
+        if (CurrentHP <= 0.0f)
         {
             //사망
-            CurrentHP = 0;
+            CurrentHP = 0.0f;
             return false;
         }
         //참이면 앞쪽 거짓이면 뒤쪽이 계산된다.
